refactor: SessionUtil 위치 이동
